connectors/fsr-agent-communication-bridge_1_1_0/requirements.txt
web.py==0.62
python-jose==3.3.0
arrow
cheroot==8.6.0
pyOpenSSL
